practice/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.0=pypi_0
aiohttp=3.10.5=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
attrs=24.2.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
dataclasses-json=0.6.7=pypi_0
deprecated=1.2.14=pypi_0
dirtyjson=1.0.8=pypi_0
filelock=3.16.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.9.0=pypi_0
greenlet=3.0.3=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.24.6=pypi_0
idna=3.8=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
llama-index-core=0.11.7=pypi_0
llama-index-embeddings-huggingface=0.3.1=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.22.0=pypi_0
minijinja=2.2.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.3=pypi_0
nltk=3.9.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.68=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openssl=3.0.15=h5eee18b_0
packaging=24.1=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py311h06a4308_0
pydantic=2.9.1=pypi_0
pydantic-core=2.23.3=pypi_0
python=3.11.9=h955ad1f_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.7.24=pypi_0
requests=2.32.3=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.5.1=pypi_0
scipy=1.14.1=pypi_0
sentence-transformers=3.0.1=pypi_0
setuptools=72.1.0=py311h06a4308_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.34=pypi_0
sqlite=3.45.3=h5eee18b_0
sympy=1.13.2=pypi_0
tenacity=8.5.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
torch=2.4.1=pypi_0
tqdm=4.66.5=pypi_0
transformers=4.44.2=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=pypi_0
wheel=0.43.0=py311h06a4308_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h5eee18b_1
yarl=1.11.0=pypi_0
zlib=1.2.13=h5eee18b_1
